pyproject.toml
[project]
name = "sysinit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
]
